Fix: Migrated to Metadata API + not-found page improvements
Migrated from SEO page to Metadata API, introduced different layouts for each page to differentiate server side components and client side components, improved not-found page
