+++ src/app/account/layout.tsx
@@ -1,0 +1,16 @@
+'use client';
+import * as React from 'react';
+
+import Seo from '@/components/Seo';
+
+
+export default function AccountPage() {
+  return (
+    <>
+      <Seo />
+      <main className='flex items-center justify-center'>
+        <h1>Account Settings Here</h1>
+      </main>
+    </>
+  );
+}
