+++ src/app/logout/page.tsx
@@ -1,25 +1,8 @@
-'use client';
+import { Metadata } from "next";
 
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
-
-export default function LogoutPage() {
-  const router = useRouter();
-  const signedin = true;
-
-  useEffect(() => {
-    if (signedin) {
-      // todo: request logout
-    }
-
-    router.push('/login');
-  }, [router, signedin]);
-
-  return signedin ? (
-    <div className='text-light flex h-full items-center justify-center'>
-      <h1>Logging you out...</h1>
-    </div>
-  ) : (
-    <></>
-  );
+export const metadata: Metadata = {
+  title: "Logout",
+}
+export default function Logout(){
+  return;
 }
