+++ src/app/not-found.tsx
@@ -10,8 +10,8 @@ export default function NotFoundPage() {
       <Seo templateTitle='Not Found' />
 
       <main className='h-full'>
-        <section className='h-full bg-white'>
-          <div className='layout flex h-full flex-col items-center justify-center text-center text-black'>
+        <section className='h-full'>
+          <div className='layout flex h-full flex-col items-center justify-center text-center text-light font-quicksand'>
             <RiAlarmWarningFill
               size={60}
               className='drop-shadow-glow animate-flicker text-red-500'
