+++ src/app/search/layout.tsx
@@ -1,0 +1,15 @@
+'use client';
+import * as React from 'react';
+
+import Seo from '@/components/Seo';
+
+export default function Search() {
+  return (
+    <>
+      <Seo /> 
+      <main className='flex items-center justify-center'>
+        <h1>Searching...</h1>
+      </main>
+    </>
+  );
+}
