+++ src/app/search/page.tsx
@@ -1,16 +1,7 @@
-'use client';
-import * as React from 'react';
-
-import Seo from '@/components/Seo';
-
-
-export default function Search() {
-  return (
-    <>
-      <Seo /> 
-      <main className='flex items-center justify-center'>
-        <h1>Searching...</h1>
-      </main>
-    </>
-  );
+import { Metadata } from "next"
+export const metadata: Metadata = {
+  title:"Search",
+}
+export default function Search(){
+  return;
 }
