+++ src/components/carousel/CarouselItem.tsx
@@ -1,4 +1,3 @@
-
 import clsxm from '@/lib/clsxm';
 
 import NextImage from '@/components/NextImage';
